+++ packages/integration-tests/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0-alpha.2](https://github.com/ZbayApp/monorepo/compare/integration-tests@0.2.0-alpha.0...integration-tests@0.3.0-alpha.2) (2022-02-21)
+
+**Note:** Version bump only for package integration-tests
+
+
+
+
+
+# [0.3.0-alpha.1](https://github.com/ZbayApp/monorepo/compare/integration-tests@0.2.0-alpha.0...integration-tests@0.3.0-alpha.1) (2022-02-21)
+
+**Note:** Version bump only for package integration-tests
+
+
+
+
+
+# [0.3.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/integration-tests@0.2.0-alpha.0...integration-tests@0.3.0-alpha.0) (2022-02-18)
+
+**Note:** Version bump only for package integration-tests
+
+
+
+
+
 # [0.2.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/integration-tests@0.1.0-alpha.5...integration-tests@0.2.0-alpha.0) (2022-02-09)
 
 **Note:** Version bump only for package integration-tests